Paginate Manager
Moved pagination logic out of GetFollowers and into a manager called paginateManager in common/manager.py.
Removed UserManager out of account/forms.py and added it to a separate account/manager.py.
Made paginate a PaginateMangaer of CustomUser.
