+++ account/forms.py
@@ -1,11 +1,7 @@
-from django import forms
 from django.contrib.auth import get_user_model
 from django.contrib.auth.forms import UserCreationForm
-from django.contrib.auth.models import BaseUserManager
 
-
-class UserManager(BaseUserManager):
-    pass
+from django import forms
 
 
 class UserRegistrationForm(UserCreationForm):
