+++ account/managers.py
@@ -1,0 +1,5 @@
+from django.contrib.auth.models import BaseUserManager
+
+
+class UserManager(BaseUserManager):
+    pass
